services/medical-records-service/requirements.txt
Flask
Flask-PyMongo
pymongo
requests
python-dotenv==0.19.2
flask_jwt_extended
flask-restplus